# Optimize mask merging and improve memory use