--- a/tests/test_coco_format.py
+++ b/tests/test_coco_format.py
@@ -390,11 +390,11 @@ class CocoConverterTest(TestCase):
             DatasetItem(id=1, image=np.zeros((5, 10, 3)),
                 annotations=[
                     Polygon(
-                        [3.0, 2.5, 1.0, 0.0, 3.5, 0.0, 3.0, 2.5],
+                        [1, 0, 3, 2, 3, 0, 1, 0],
                         label=3, id=4, group=4,
                         attributes={ 'is_crowd': False }),
                     Polygon(
-                        [5.0, 3.5, 4.5, 0.0, 8.0, 0.0, 5.0, 3.5],
+                        [5, 0, 5, 3, 8, 0, 5, 0],
                         label=3, id=4, group=4,
                         attributes={ 'is_crowd': False }),
                 ], attributes={'id': 1}